tests/e2e/pyproject.toml
[project]
name = "e2e"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faker>=37.12.0",
    "pytest-playwright>=0.7.1",
]
